+++ tests/test_molecule.py
@@ -1,5 +1,5 @@
-from contextlib import nullcontext
-from typing import TYPE_CHECKING, ClassVar, ContextManager
+from contextlib import AbstractContextManager, nullcontext
+from typing import TYPE_CHECKING, ClassVar
 
 import pytest
 from MDAnalysis import SelectionError
@@ -64,10 +64,22 @@ class TestMolecule(pytest.BaseTestMixinRDKitMol):  # type: ignore[name-defined]
     def test_n_residues(self, mol: Molecule) -> None:
         assert mol.n_residues == mol.residues.n_residues
 
+    def test_large_selection_automatically_uses_segid(
+        self, water_u: "Universe"
+    ) -> None:
+        """
+        If more segids than chains are present, Molecule should switch to segindex.
+        """
+        water = water_u.select_atoms("resname TIP3 and byres around 6 protein")
+        water_mol = Molecule.from_mda(water, NoImplicit=False)
+        assert water_mol["TIP34.3"]
+        # would have overwritten TIP34.3 if using chain as there's only chain X
+        assert water_mol["TIP34.4"]
+
 
 class SupplierBase:
     resid = ResidueId("UNL", 1, "")
-    slice_behavior: ClassVar[ContextManager] = nullcontext()
+    slice_behavior: ClassVar[AbstractContextManager] = nullcontext()
 
     def test_len(self, suppl: "Sequence[Molecule]") -> None:
         assert len(suppl) == 9